[add] Impliment logging and add main.py script
Before this commit, I implement logging by the function "print".
But print is not recommented for writing log because if we use print for logging, anyone does not understand that the output is the log for debug or the true output of the function immidiately.
So I replace print function by the logging function.
But the logging is not clearly implemented, so I plane to fix the structure in the nearly future.